+++ CHANGES.md
@@ -11,7 +11,7 @@ Apollo Java 2.4.0
 * [Add more observability in apollo config client](https://github.com/apolloconfig/apollo-java/pull/74)
 * [Feature Support Kubernetes ConfigMap cache for Apollo java, golang client](https://github.com/apolloconfig/apollo-java/pull/79)
 * [Feature support pulling configuration information from multiple AppIds](https://github.com/apolloconfig/apollo-java/pull/70)
-
+* [Fix monitor arg cause npe](https://github.com/apolloconfig/apollo-java/pull/86)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/4?closed=1)